adding new jwt
