+++ src/js/jwt/logger.js
@@ -1,0 +1,9 @@
+const pub = {};
+
+module.exports = (fileName) => {
+    return (msg) => {  
+        if (window.console) {
+            window.console.log(`[jwt][${fileName}] ${msg}`);
+        }
+    };
+}
