+++ src/js/jwt/utils.js
@@ -1,0 +1,9 @@
+const pub = {};
+
+pub.exposeTest = (priv) => {
+    if (window.NODE_ENV === 'test') {
+        module.exports.priv = priv;
+    }
+};
+
+module.exports = pub;
